add: 추천 채널, 기존 채널 컴포넌트 분리
